--- app/src/main/java/se/grupp1/antonsskafferi/AdminActivity.java
@@ -1,42 +1,0 @@
-package se.grupp1.antonsskafferi;
-
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.View;
-
-public class AdminActivity extends AppCompatActivity
-{
-    @Override
-    protected void onCreate(Bundle savedInstanceState)
-    {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_admin);
-
-        setTitle("Admin");
-
-
-        //Kopiera dessa två rader för att lägga en tillbaka-pil uppe till vänster
-        ActionBar actionBar = getSupportActionBar();
-        actionBar.setDisplayHomeAsUpEnabled(true);
-
-        findViewById(R.id.editLunchButton).setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View v)
-            {
-                startActivity(new Intent(getApplicationContext(), EditLunchActivity.class));
-            }
-        });
-    }
-
-    public boolean onOptionsItemSelected(MenuItem item)
-    {
-        Intent myIntent = new Intent(getApplicationContext(), MainActivity.class);
-        startActivity(myIntent);
-        return true;
-    }
-}
