--- app/src/main/java/se/grupp1/antonsskafferi/ScheduleActivity.java
@@ -1,19 +1,0 @@
-package se.grupp1.antonsskafferi;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-import android.widget.CalendarView;
-
-public class ScheduleActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_schedule);
-        CalendarView calendar = (CalendarView) findViewById(R.id.calendar); // get the reference of CalendarView
-        calendar.setDate(System.currentTimeMillis()); // set selected date 22 May 2016 in milliseconds
-        calendar.setShowWeekNumber(true);
-
-    }
-}
