+++ app/src/main/res/menu/navbar_menu.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_schedule"
+        android:title="Schedule" />
+    <item
+        android:id="@+id/navigation_admin"
+        android:title="Admin" />
+</menu>
